% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capital_V_income_eu}
\alias{capital_V_income_eu}
\title{The capital/income ratio in Europe, 1870-2010.}
\format{A data frame with 15 rows and 4 variables:
\describe{
  \item{Year}{date, only year is accurate}
  \item{Numeric values}{private capital / national income ratio}
  ...
}}
\source{
\url{http://piketty.pse.ens.fr/en/capital21c2}
}
\usage{
capital_V_income_eu
}
\description{
Time series data containing the Private capital / national income ratio for Germany, France and Britain.
}
\keyword{datasets}
