% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decile_share_us}
\alias{decile_share_us}
\title{The top decile income share in the United States.}
\format{A data frame with 101 rows and 2 variables:
\describe{
  \item{Year}{date, only year is accurate}
  \item{Percentage}{private capital / national income ratio}
}}
\source{
\url{http://piketty.pse.ens.fr/en/capital21c2}
}
\usage{
decile_share_us
}
\description{
Time series data (1910-2010) The top decile income share in the United States (included capital gains).
}
\keyword{datasets}
