% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter1figures.R
\name{get_figures}
\alias{get_figures}
\title{Retrieve all figures from a chapter in "Capital in the Twenty-First Century" by Thomas Piketty}
\usage{
get_figures(chapter)
}
\arguments{
\item{chapter}{numeric: chapter number.}
}
\value{
All figures in chapter \code{chapter}.

Figure plotting relies on dplyr, ggplot2, tidyr and scales
}
\description{
Retrieve all figures from a chapter in "Capital in the Twenty-First Century" by Thomas Piketty
}
\examples{
get_figures(chapter=0)
FI.1
FI.2

}
